Adding link to home from the logo and moving the style to a separate file
