+++ ui/src/components/Footer/Footer.jsx
@@ -1,36 +1,7 @@
 import { Box, Link } from '@mui/material';
-import makeStyles from '@mui/styles/makeStyles';
 import React from 'react';
 
-const useStyles = makeStyles((theme) => ({
-  rootBox: {
-    [theme.breakpoints.down('md')]: {
-      justifyContent: 'center',
-    },
-  },
-  footerNav: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    justifyContent: 'center',
-    marginRight: 'auto',
-    marginLeft: theme.spacing(3),
-    marginBottom: theme.spacing(0),
-
-    [theme.breakpoints.down('md')]: {
-      width: '100%',
-      marginLeft: 'auto',
-      marginTop: theme.spacing(3),
-      marginBottom: theme.spacing(2),
-    },
-  },
-  footerLink: {
-    marginLeft: theme.spacing(3),
-    marginRight: theme.spacing(3),
-    [theme.breakpoints.down('md')]: {
-      marginBottom: theme.spacing(2),
-    },
-  },
-}));
+import { useStyles } from './Footer.styles';
 
 export default function Footer() {
   const classes = useStyles();
@@ -41,7 +12,6 @@ export default function Footer() {
         sx={{
           paddingRight: '16px',
           paddingLeft: '16px',
-          backgroundColor: '#F7F7F7',
           color: '#000',
           borderTop: '1px solid #E1E1E1',
         }}
