+++ ui/src/components/Footer/Footer.styles.js
@@ -1,0 +1,33 @@
+import makeStyles from '@mui/styles/makeStyles';
+
+const styles = (theme) => ({
+  rootBox: {
+    [theme.breakpoints.down('md')]: {
+      justifyContent: 'center',
+    },
+  },
+  footerNav: {
+    display: 'flex',
+    flexWrap: 'wrap',
+    justifyContent: 'center',
+    marginRight: 'auto',
+    marginLeft: theme.spacing(3),
+    marginBottom: theme.spacing(0),
+
+    [theme.breakpoints.down('md')]: {
+      width: '100%',
+      marginLeft: 'auto',
+      marginTop: theme.spacing(3),
+      marginBottom: theme.spacing(2),
+    },
+  },
+  footerLink: {
+    marginLeft: theme.spacing(3),
+    marginRight: theme.spacing(3),
+    [theme.breakpoints.down('md')]: {
+      marginBottom: theme.spacing(2),
+    },
+  },
+});
+
+export const useStyles = makeStyles(styles, { name: 'Footer' });
